ci: rework laberler/pr-info workflows
This folds the `pr-info` job into the `labeler` workflow, allowing the
single workflow to be the new starting point for the rest of the CI
jobs.

Because we need the information gathered in the `pr-info` job, we need
to save it as an artifact to be used in other workflows.

Signed-off-by: Micah Abbott <miabbott@redhat.com>